Move quick_button_preferences to proto package
