+++ app/src/test/java/com/jkuester/unlauncher/datasource/DataSourceModuleTest.kt
@@ -3,8 +3,8 @@ package com.jkuester.unlauncher.datasource
 import android.content.Context
 import androidx.datastore.core.DataStore
 import com.jkuester.unlauncher.datastore.CorePreferences
-import com.jkuester.unlauncher.datastore.QuickButtonPreferences
 import com.jkuester.unlauncher.datastore.UnlauncherApps
+import com.jkuester.unlauncher.datastore.proto.QuickButtonPreferences
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
